Split into modules for easier maintainability
Even for simple codebases, unlike Babel and TypeScript, it's important to
write well-designed code
